+++ alicloud/resource_alicloud_disk.go
@@ -140,7 +140,7 @@ func resourceAliyunDiskRead(d *schema.ResourceData, meta interface{}) error {
 	})
 
 	if err != nil {
-		if notFoundError(err) {
+		if NotFoundError(err) {
 			d.SetId("")
 			return nil
 		}