+++ alicloud/resource_alicloud_nat_gateway_test.go
@@ -158,18 +158,16 @@ func testAccCheckNatGatewayDestroy(s *terraform.State) error {
 
 		// Try to find the Nat gateway
 		instance, err := client.DescribeNatGateway(rs.Primary.ID)
-
 		if instance != nil {
 			return fmt.Errorf("Nat gateway still exist")
 		}
 
-		if err != nil {
-			// Verify the error is what we want
-			e, _ := err.(*common.Error)
-
-			if !notFoundError(e) {
-				return err
+		if err != nil && !NotFoundError(err) {
+			if e, ok := err.(*common.Error); ok && (e.Code == AliyunGoClientFailure || e.StatusCode == -1) {
+				return nil
 			}
+			// Verify the error is what we want
+			return err
 		}
 
 	}
@@ -190,7 +188,7 @@ resource "alicloud_vpc" "foo" {
 resource "alicloud_vswitch" "foo" {
 	vpc_id = "${alicloud_vpc.foo.id}"
 	cidr_block = "172.16.0.0/21"
-	availability_zone = "${data.alicloud_zones.default.zones.2.id}"
+	availability_zone = "${data.alicloud_zones.default.zones.0.id}"
 }
 
 resource "alicloud_nat_gateway" "foo" {
@@ -200,19 +198,19 @@ resource "alicloud_nat_gateway" "foo" {
 	bandwidth_packages = [{
 	  ip_count = 1
 	  bandwidth = 5
-	  zone = "${data.alicloud_zones.default.zones.2.id}"
+	  zone = "${data.alicloud_zones.default.zones.0.id}"
 	}, {
 	  ip_count = 2
 	  bandwidth = 6
-	  zone = "${data.alicloud_zones.default.zones.2.id}"
+	  zone = "${data.alicloud_zones.default.zones.0.id}"
 	}, {
 	  ip_count = 3
 	  bandwidth = 7
-	  zone = "${data.alicloud_zones.default.zones.2.id}"
+	  zone = "${data.alicloud_zones.default.zones.0.id}"
 	}, {
 	  ip_count = 1
 	  bandwidth = 8
-	  zone = "${data.alicloud_zones.default.zones.2.id}"
+	  zone = "${data.alicloud_zones.default.zones.0.id}"
 	}]
 	depends_on = [
     	"alicloud_vswitch.foo"]