+++ alicloud/resource_alicloud_vpc_test.go
@@ -70,6 +70,34 @@ func TestAccAlicloudVpc_update(t *testing.T) {
 	})
 }
 
+func TestAccAlicloudVpc_multi(t *testing.T) {
+	var vpc ecs.VpcSetType
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckVpcDestroy,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccVpcConfigMulti,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVpcExists("alicloud_vpc.bar_1", &vpc),
+					resource.TestCheckResourceAttr(
+						"alicloud_vpc.bar_1", "cidr_block", "172.16.0.0/12"),
+					testAccCheckVpcExists("alicloud_vpc.bar_2", &vpc),
+					resource.TestCheckResourceAttr(
+						"alicloud_vpc.bar_2", "cidr_block", "192.168.0.0/16"),
+					testAccCheckVpcExists("alicloud_vpc.bar_3", &vpc),
+					resource.TestCheckResourceAttr(
+						"alicloud_vpc.bar_3", "cidr_block", "10.1.0.0/21"),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckVpcExists(n string, vpc *ecs.VpcSetType) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
@@ -136,5 +164,20 @@ const testAccVpcConfigUpdate = `
 resource "alicloud_vpc" "foo" {
 	cidr_block = "172.16.0.0/12"
 	name = "tf_test_bar"
+}
+`
+
+const testAccVpcConfigMulti = `
+resource "alicloud_vpc" "bar_1" {
+	cidr_block = "172.16.0.0/12"
+	name = "tf_test_bar_1"
+}
+resource "alicloud_vpc" "bar_2" {
+	cidr_block = "192.168.0.0/16"
+	name = "tf_test_bar_2"
+}
+resource "alicloud_vpc" "bar_3" {
+	cidr_block = "10.1.0.0/21"
+	name = "tf_test_bar_3"
 }
 `